Hold balances as BigDecimal instead of double
Balances were double, so ordinary arithmetic produced values like
749.6610000000001: most decimal amounts have no exact binary representation,
and the error compounds across trades until totals stop reconciling.

User, P2PListing and P2POrder now use BigDecimal on numeric(36,18) columns -
room for cedi balances in the billions, and 18 decimal places, which covers a
satoshi at 8 and wei at 18. ListingResponse and OrderResponse carried Double
and were re-rounding the values on the way back out; they match now.

The existing columns were migrated with a cast through text
(cedi_balance::text::numeric) so each row keeps the value that was actually
stored rather than gaining precision the double never had.

Changing the field type turned every call site into a compile error, which is
how all eighty were found rather than guessed at. Several were doing their own
per-asset switch over balances - P2PServiceImpl, P2PController,
SpotTradingServiceImpl and AssetServiceImpl - and they recognised different
sets of assets: P2PController knew only BTC and ETH, P2PServiceImpl did not
know GHS. They delegate to User.getBalanceFor/setBalanceFor now, so the
supported list is stated once.

Paystack's crypto deposit divided amount by price in double. It divides in
BigDecimal with an explicit RoundingMode.HALF_UP at 18 places, so the rounding
is a decision rather than whatever the hardware did.

Also removes six hand-written balance getters on User that shadowed Lombok's
@Data. Keeping them in step with the field types by hand is how they came to
return the wrong type.

Verified against the running backend, every money path:

  buy      749.661000000000100000 - 250.92 = 498.741000000000100000 exactly,
           and 2.1 + 0.1 = 2.2, where double gives 2.2000000000000002
  convert  100 USDT -> 0.001283715 BTC, both sides exact; overdraft still
           refused
  spot     0.001 BTC at 500: BTC nets to zero across the two accounts and the
           USDT difference is exactly the 0.00075 in maker and taker fees
  p2p      3 USDT at 15.45: 46.35 GHS moved, both sides exact

The 000100000 tail on that USD balance is a value stored before the migration;
the arithmetic over it is now exact, but historical noise is preserved rather
than invented away.

+++ backend/src/main/java/project/bitby/bitby/dto/ListingResponse.java
@@ -1,5 +1,6 @@
 package project.bitby.bitby.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,8 +15,8 @@ public class ListingResponse {
     private Long id;
     private String sellerUsername;
     private String assetSymbol;
-    private Double amount;
-    private Double pricePerUnit;
+    private BigDecimal amount;
+    private BigDecimal pricePerUnit;
     private String currency;
     private P2PListing.ListingStatus status;
     private LocalDateTime createdAt;
