+++ backend/src/main/java/project/bitby/bitby/dto/OrderResponse.java
@@ -1,5 +1,6 @@
 package project.bitby.bitby.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -17,9 +18,9 @@ public class OrderResponse {
     private String buyerUsername;
     private String sellerUsername;
     private String assetSymbol;
-    private Double amount;
-    private Double totalPrice;
-    private Double pricePerUnit;
+    private BigDecimal amount;
+    private BigDecimal totalPrice;
+    private BigDecimal pricePerUnit;
     private String currency;
     private String paymentMethod;
     private P2POrder.OrderType orderType;
